+++ src/toolbox/css/_buttons.scss
@@ -20,4 +20,7 @@
         float: inline-end;
         font-size: 30px;
     }
+    &-darkmode {
+        font-size: 20px;
+    }
 }
