+++ src/toolbox/css/_var.scss
@@ -4,6 +4,12 @@ $pastel: #ffebcd;
 $pastel-secondary: #f5f5dc;
 $pastel-blur: #fffaf0;
 $greenish: #008080;
+// darkmode
+$darkmode_pastel: #efc37b;
+$darkmode_pastel-secondary: rgb(163, 221, 192);
+$darkmode_pastel-blur: #303030;
+$darkmode_greenish: #013939;
+// messages
 $success-green: #4F8A10;
 $field-success-green:#DFF2BF;
 $error-red: #D8000C;
