+++ src/toolbox/partials/footer.php
@@ -1,6 +1,7 @@
-<div class="cut-off"></div>
-<script src="../js/infoButtonToggle.js"></script>
-<script src="../js/InputFiieldToggle.js"></script>
+<div id="cut-off" class="cut-off"></div>
+<script src="./src/toolbox/js/infoButtonToggle.js"></script>
+<script src="./src/toolbox/js/InputFiieldToggle.js"></script>
+<script src="./src/toolbox/js/enableDarkmode.js"></script>
 <small>v1.1</small>
 
 <script src="https://kit.fontawesome.com/ee02357a59.js"></script>
