+++ src/toolbox/partials/header.php
@@ -7,7 +7,7 @@
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <link rel="icon" href="http://example.com/favicon.png">
-   <link rel="stylesheet" href="../css/main.css">
+   <link rel="stylesheet" href="./src/toolbox/css/main.css">
    <title><?php echo $host ?></title>
 
 </head>
