+++ src/toolbox/partials/nav.php
@@ -1,0 +1,13 @@
+<div id="heading" class="heading">
+
+   <div class="logo-container">
+      <a class="logo fab fa-dev fa-3x"></a>
+      <a href="/" class="logo-txt"><?php echo $SLASH . $host  ?>
+      </a>
+   </div>
+   <div class="options">
+      <i id="darkmodeToggleOn" class="fas fa-toggle-on btn-darkmode hide"></i>
+      <i id="darkmodeToggleOff" class="fas fa-toggle-off btn-darkmode"></i>
+   </div>
+
+</div>
